user needs to access only his own tasks

+++ src/tasks/tasks.controller.ts
@@ -3,6 +3,7 @@ import {
   Body,
   Controller,
   Delete,
+  ForbiddenException,
   Get,
   HttpCode,
   HttpStatus,
@@ -23,7 +24,6 @@ import { CreateTaskLabelDto } from './create-task-label.dto';
 import { FindTaskParams } from './find-task.params';
 import { PaginationParams } from '../common/pagination.params';
 import { PaginationResponse } from 'src/common/pagination.response';
-import type { AuthRequest } from 'src/users/auth.request';
 import { CurrentUserId } from '../users/decorators/current-user-id.decorator';
 
 @Controller('tasks')
@@ -48,8 +48,13 @@ export class TasksController {
   }
 
   @Get('/:id')
-  public async findOne(@Param() params: FindOneParams): Promise<Task> {
-    return await this.findOneOrFail(params.id);
+  public async findOne(
+    @Param() params: FindOneParams,
+    @CurrentUserId() userId: string,
+  ): Promise<Task> {
+    const task = await this.findOneOrFail(params.id);
+    this.checkTaskOwnership(task, userId);
+    return task;
   }
 
   @Post()
@@ -68,10 +73,11 @@ export class TasksController {
   public async updateTask(
     @Param() params: FindOneParams,
     @Body() updateTaskDto: UpdateTaskDto,
+    @CurrentUserId() userId: string,
   ): Promise<Task> {
     try {
       const task = await this.findOneOrFail(params.id);
-
+      this.checkTaskOwnership(task, userId);
       return await this.tasksService.updateTask(task, updateTaskDto);
     } catch (error) {
       if (error instanceof WrongTaskStatusException) {
@@ -83,8 +89,12 @@ export class TasksController {
 
   @Delete('/:id')
   @HttpCode(HttpStatus.NO_CONTENT)
-  public async deleteTask(@Param() params: FindOneParams): Promise<void> {
+  public async deleteTask(
+    @Param() params: FindOneParams,
+    @CurrentUserId() userId: string,
+  ): Promise<void> {
     const task = await this.findOneOrFail(params.id);
+    this.checkTaskOwnership(task, userId);
     await this.tasksService.deleteTask(task);
   }
 
@@ -102,8 +112,10 @@ export class TasksController {
   async addLabels(
     @Param() { id }: FindOneParams,
     @Body() labels: CreateTaskLabelDto[],
+    @CurrentUserId() userId: string,
   ): Promise<Task> {
     const task = await this.findOneOrFail(id);
+    this.checkTaskOwnership(task, userId);
     return this.tasksService.addLabels(task, labels);
   }
 
@@ -112,8 +124,16 @@ export class TasksController {
   async removeLabels(
     @Param() { id }: FindOneParams,
     @Body() labelsToRemove: string[],
+    @CurrentUserId() userId: string,
   ): Promise<void> {
     const task = await this.findOneOrFail(id);
+    this.checkTaskOwnership(task, userId);
     await this.tasksService.removeLabels(task, labelsToRemove);
+  }
+
+  private checkTaskOwnership(task: Task, userId: string): void {
+    if (task.userId != userId) {
+      throw new ForbiddenException('This task does not belog to this user');
+    }
   }
 }
